Cause RecurrenceIterator to respect the timezone of the event that is recurring.
The UNTIL value for the RRULE of an event is being parsed without a
timezone, thus all UNTIL values are in UTC timezone, even if that is not
what is indicated.

When the UNTIL value comes from a valid ICS feed that specifies a
timezone, but does not end with 'Z', the timezone of the event should
be used.

If the UNTIL value ends in "Z", it will still be interpreted as a UTC
value, and work the same way as current functionality.