Validator upgrades.
This comment adds two new options to the validator:

1. PROFILE_CARDDAV
2. PROFILE_CALDAV

Passing these options will tell the validator to assume that the
vcard/icalendar object needs to be validated specifically according to
the rules of caldav/carddav.

For vCards specifically this means that the UID is required for CardDAV,
but we only emit a warning without PROFILE_CARDDAV.

For iCalendar there's a bunch of new rules that we will eventually
integrate into sabredav.

This fixes #176.