adjust unit test settings for time limits
Some tests were testing a bug that caused an infinite loop.

Annotate those tests with large, small annotations.

Turn on enforceTimeLimit, failOnWarning and failOnRisky so that the
annotations are enforced when unit tests are run.

Add phpunit/php-invoker to the composer require-dev becaause this is
required to make the enforceTimeLimit setting effective.