Ignore multiple same parameter-values
Sometimes it seems like parameters can be added multiple times with
the same value. Besides not making any sense (and partially being
not allowed in the first place) this will also possibly break the
further process of parsing an icalendar file.

This modification adds a shortcut when the last parameter-value and the
current parameter value are exactly the same. In that case the new
value is ignored and only the old value will be used.

The use-case - as documented in the test - is that there seem to be
icalendar files that contain the following content:

    DTSTART;VALUE=DATE;VALUE=DATE:20220612

While that seems to violate the RFC it is not seen as a violation in
the icalendar validator at https://icalendar.org/validator.html and
as it can also be imported by other calendaring systems it looks like
this should at least not break the parser.