Fix: Remove X-ADDRESSBOOKSERVER-KIND=individual property in convert
When converting to VCard v4, only X-ADDRESSBOOKSERVER-KIND=GROUP is
converted to KIND=group currently. X-ADDRESSBOOKSERVER-KIND=INDIVIDUAL
is kept in the vcard, although KIND=individual would be correct. Since
this is the default, the property can also be dropped, as is done by
vobject in the opposite direction (converting v4 to v3). Therefore, this
fix will also drop the X-ADDRESSBOOKSERVER-KIND=INDIVIDUAL property.

Test existing test was broken because it used the output of the v4->v3
conversion as input for the v3->v4 conversion, but since the property is
dropped during the first conversion there is no explicit
X-ADDRESSBOOKSERVER-KIND property in the input of the v3->v4 conversion.
Test fixed as well.