Allow unknown value data types for VALUE
Currently value-types that are not known will cause the parser to throw
an InvalidDataException. In RFC5545, Section 3.2.20 is described though
that the allowed values for the "VALUE" data-type can include x-names and
iana-tokens. These can be any string that might - or might not - be
understood by the parser. The description clearly states that
"Applications MUST preserve the value data for x-name and iana-token
value that they don't recognize without attempting to interpret or parse
the value data"

This means that the content of the "VALUE" part should not be
interpreted at all the moment the parser doesn't find a match in the
corresponding mapping table.

But it should also not hard-fail when someone sets a VALUE that might
not be understood. In such a case the default should be used instead.