Add tests for skipping unknown parameters
THis adds a test to assure that an unknown value is skipped. The value
`DATETIME` is not known. It should be `DATE-TIME`. Adding the value of
DATETIME previopusly resulted in an Error. This now tests that it
actually works and that the unknown parameter is actually also kept as
specified by the RFC